Finished normalizing file names
